requirements.txt
Django==4.1
djangorestframework==3.14.0
Pillow==9.5.0

